+++ packages/core/field-text/src/FieldText.js
@@ -37,7 +37,7 @@ export default class FieldText extends Component<FieldTextProps, State> {
         {...this.props}
         value={this.state.value}
         onChange={this.handleOnChange}
-        ref={fieldRef => {
+        innerRef={(fieldRef: ?HTMLInputElement) => {
           this.input = fieldRef;
         }}
       />