+++ packages/editor/editor-common/src/schema/marks/index.ts
@@ -1,17 +1,29 @@
-export { em } from './em';
-export { code } from './code';
-export { strike } from './strike';
-export { strong } from './strong';
-export { underline } from './underline';
-export { link, LinkAttributes, toJSON as linkToJSON } from './link';
+export { em, EmDefinition } from './em';
+export { code, CodeDefinition } from './code';
+export { strike, StrikeDefinition } from './strike';
+export { strong, StrongDefinition } from './strong';
+export { underline, UnderlineDefinition } from './underline';
+export {
+  link,
+  LinkAttributes,
+  LinkDefinition,
+  toJSON as linkToJSON,
+} from './link';
 export { emojiQuery } from './emoji-query';
 export { mentionQuery } from './mention-query';
 export { typeAheadQuery } from './type-ahead-query';
-export { subsup } from './subsup';
-export { textColor, colorPalette, borderColorPalette } from './text-color';
+export { subsup, SubSupDefinition, SubSupAttributes } from './subsup';
+export {
+  textColor,
+  colorPalette,
+  borderColorPalette,
+  TextColorDefinition,
+  TextColorAttributes,
+} from './text-color';
 export { confluenceInlineComment } from './confluence-inline-comment';
 export {
   action,
   Action as ActionMarkAction,
-  Attributes as ActionMarkAttributes,
+  ActionMarkAttributes,
+  ActionDefinition,
 } from './action';