+++ packages/media/media-viewer/__tests__/newgen/viewers/video.spec.tsx
@@ -4,14 +4,14 @@ const constructAuthTokenUrlSpy = jest.spyOn(util, 'constructAuthTokenUrl');
 import * as React from 'react';
 import { mount } from 'enzyme';
 import Button from '@atlaskit/button';
-import { Stubs } from '../../_stubs';
-import { Subject } from 'rxjs/Subject';
-import { FileItem } from '@atlaskit/media-core';
+import { createContext } from '../../_stubs';
+import { FileItem, Auth } from '@atlaskit/media-core';
 import { VideoViewer, Props } from '../../../src/newgen/viewers/video';
 import { Video } from '../../../src/newgen/styled';
 import Spinner from '@atlaskit/spinner';
 import { ErrorMessage } from '../../../src/newgen/styled';
 import { CustomVideo } from '../../../src/newgen/viewers/video/customVideo';
+import { awaitError } from '@atlaskit/media-test-helpers';
 
 const token = 'some-token';
 const clientId = 'some-client-id';
@@ -33,36 +33,17 @@ const videoItem: FileItem = {
   },
 };
 
-function createContext(authPromise) {
-  const serviceHost = 'some-service-host';
-  const authProvider = jest.fn().mockReturnValue(authPromise);
-  const contextConfig = {
-    serviceHost,
-    authProvider,
-  };
-  return Stubs.context(
-    contextConfig,
-    undefined,
-    Stubs.mediaItemProvider(new Subject<FileItem>()),
-  ) as any;
-}
-
-function createFixture(authPromise, props?: Partial<Props>) {
-  const context = createContext(authPromise);
+function createFixture(authPromise: Promise<Auth>, props?: Partial<Props>) {
+  const context = createContext({ authPromise });
   const el = mount(
-    <VideoViewer context={context} item={videoItem} {...props} />,
+    <VideoViewer
+      context={context}
+      item={videoItem}
+      {...props}
+      previewCount={0}
+    />,
   );
   return { context, el };
-}
-
-async function awaitError(response, expectedMessage) {
-  try {
-    await response;
-  } catch (err) {
-    if (err.message !== expectedMessage) {
-      throw err;
-    }
-  }
 }
 
 describe('Video viewer', () => {
@@ -73,7 +54,7 @@ describe('Video viewer', () => {
   it('assigns a src for videos when successful', async () => {
     const authPromise = Promise.resolve({ token, clientId });
     const { el } = createFixture(authPromise);
-    await el.instance()['init']();
+    await (el as any).instance()['init']();
     el.update();
     expect(el.find(Video).prop('src')).toEqual(
       'some-service-host/video?client=some-client-id&token=some-token',
@@ -81,7 +62,7 @@ describe('Video viewer', () => {
   });
 
   it('shows spinner when pending', async () => {
-    const authPromise = new Promise(() => {});
+    const authPromise: any = new Promise(() => {});
     const { el } = createFixture(authPromise);
     el.update();
     expect(el.find(Spinner)).toHaveLength(1);
@@ -99,7 +80,7 @@ describe('Video viewer', () => {
     const collectionName = 'some-collection';
     const authPromise = Promise.resolve({ token, clientId });
     const { el } = createFixture(authPromise, { collectionName });
-    await el.instance()['init']();
+    await (el as any).instance()['init']();
     el.update();
     expect(constructAuthTokenUrlSpy.mock.calls[0][2]).toEqual(collectionName);
   });
@@ -110,7 +91,7 @@ describe('Video viewer', () => {
       featureFlags: { customVideoPlayer: true },
     });
 
-    await el.instance()['init']();
+    await (el as any).instance()['init']();
     el.update();
 
     expect(el.find(CustomVideo)).toHaveLength(1);
@@ -125,7 +106,7 @@ describe('Video viewer', () => {
       featureFlags: { customVideoPlayer: true },
     });
 
-    await el.instance()['init']();
+    await (el as any).instance()['init']();
     el.update();
     expect(el.state('isHDActive')).toBeFalsy();
     el
@@ -133,5 +114,50 @@ describe('Video viewer', () => {
       .at(1)
       .simulate('click');
     expect(el.state('isHDActive')).toBeTruthy();
+  });
+
+  describe('AutoPlay', () => {
+    async function createAutoPlayFixture(
+      previewCount: number,
+      isCustomVideoPlayer: boolean,
+    ) {
+      const authPromise = Promise.resolve({ token, clientId });
+      const context = createContext({ authPromise });
+      const el = mount(
+        <VideoViewer
+          context={context}
+          previewCount={previewCount}
+          item={videoItem}
+          featureFlags={{ customVideoPlayer: isCustomVideoPlayer }}
+        />,
+      );
+      await (el as any).instance()['init']();
+      el.update();
+      return el;
+    }
+
+    it('should auto play custom video viewer when it is the first preview', async () => {
+      const el = await createAutoPlayFixture(0, true);
+      expect(el.find(CustomVideo)).toHaveLength(1);
+      expect(el.find({ autoPlay: true })).toHaveLength(2);
+    });
+
+    it('should not auto play custom video viewer when it is not the first preview', async () => {
+      const el = await createAutoPlayFixture(1, true);
+      expect(el.find(CustomVideo)).toHaveLength(1);
+      expect(el.find({ autoPlay: true })).toHaveLength(0);
+    });
+
+    it('should auto play native video viewer when it is the first preview', async () => {
+      const el = await createAutoPlayFixture(0, false);
+      expect(el.find(CustomVideo)).toHaveLength(0);
+      expect(el.find({ autoPlay: true })).toHaveLength(2);
+    });
+
+    it('should not auto play native video viewer when it is not the first preview', async () => {
+      const el = await createAutoPlayFixture(1, false);
+      expect(el.find(CustomVideo)).toHaveLength(0);
+      expect(el.find({ autoPlay: true })).toHaveLength(0);
+    });
   });
 });