+++ packages/people-and-teams/profilecard/examples/helper/util.js
@@ -1,0 +1,16 @@
+// @flow
+import { getMockProfileClient as getMockProfileClientUtil } from '../../mock-helpers';
+import { AkProfileClient, modifyResponse } from '../../src';
+
+const MockProfileClient = getMockProfileClientUtil(
+  AkProfileClient,
+  modifyResponse,
+);
+
+export const getMockProfileClient = (cacheSize: number, cacheMaxAge: number) =>
+  new MockProfileClient({
+    cacheSize,
+    cacheMaxAge,
+  });
+
+export default null;