+++ packages/search/global-search/src/components/confluence/ConfluenceSearchResults.tsx
@@ -12,7 +12,18 @@ import {
   searchPeopleItem,
   take,
   isEmpty,
+  getConfluenceAdvancedSearchLink,
 } from '../SearchResultsUtil';
+import AnalyticsEventFiredOnMount from '../analytics/AnalyticsEventFiredOnMount';
+import { buildScreenEvent, Screen } from '../../util/analytics-util';
+import NoRecentActivity from '../NoRecentActivity';
+
+export const MAX_PAGES_BLOGS_ATTACHMENTS = 8;
+export const MAX_SPACES = 3;
+export const MAX_PEOPLE = 3;
+
+let preQueryScreenCounter = 0;
+let postQueryScreenCounter = 0;
 
 const renderObjectsGroup = (title: string, results: Result[], query: string) =>
   results.length > 0 ? (
@@ -95,6 +106,7 @@ export interface Props {
   objectResults: Result[];
   spaceResults: Result[];
   peopleResults: Result[];
+  searchSessionId: string;
 }
 
 export default function searchResults(props: Props) {
@@ -109,6 +121,7 @@ export default function searchResults(props: Props) {
     objectResults,
     spaceResults,
     peopleResults,
+    searchSessionId,
   } = props;
 
   if (isLoading) {
@@ -120,20 +133,58 @@ export default function searchResults(props: Props) {
   }
 
   if (query.length === 0) {
+    if (
+      [
+        recentlyInteractedPeople,
+        recentlyViewedPages,
+        recentlyViewedSpaces,
+      ].every(isEmpty)
+    ) {
+      return (
+        <NoRecentActivity
+          advancedSearchUrl={getConfluenceAdvancedSearchLink()}
+        />
+      );
+    }
     // TODO: insert error state here if the recent results are empty.
+    if (
+      [
+        recentlyInteractedPeople,
+        recentlyViewedPages,
+        recentlyViewedSpaces,
+      ].every(isEmpty)
+    ) {
+      return null;
+    }
+
     return [
       renderObjectsGroup(
         'Recent pages and blogs',
-        take(recentlyViewedPages, 8),
+        take(recentlyViewedPages, MAX_PAGES_BLOGS_ATTACHMENTS),
         query,
       ),
-      renderSpacesGroup('Recent spaces', take(recentlyViewedSpaces, 3), query),
+      renderSpacesGroup(
+        'Recent spaces',
+        take(recentlyViewedSpaces, MAX_SPACES),
+        query,
+      ),
       renderPeopleGroup(
         'Recently worked with',
-        take(recentlyInteractedPeople, 3),
+        take(recentlyInteractedPeople, MAX_PEOPLE),
         query,
       ),
       renderAdvancedSearchGroup(query),
+      <AnalyticsEventFiredOnMount
+        key="preQueryScreenEvent"
+        onEventFired={() => preQueryScreenCounter++}
+        payloadProvider={() =>
+          buildScreenEvent(
+            Screen.PRE_QUERY,
+            preQueryScreenCounter,
+            searchSessionId,
+          )
+        }
+      />,
     ];
   }
 
@@ -144,11 +195,23 @@ export default function searchResults(props: Props) {
   return [
     renderObjectsGroup(
       'Pages, blogs and attachments',
-      take(objectResults, 8),
+      take(objectResults, MAX_PAGES_BLOGS_ATTACHMENTS),
       query,
     ),
-    renderSpacesGroup('Spaces', take(spaceResults, 3), query),
-    renderPeopleGroup('People', take(peopleResults, 3), query),
+    renderSpacesGroup('Spaces', take(spaceResults, MAX_SPACES), query),
+    renderPeopleGroup('People', take(peopleResults, MAX_PEOPLE), query),
     renderAdvancedSearchGroup(query),
+
+    <AnalyticsEventFiredOnMount
+      key="postQueryScreenEvent"
+      onEventFired={() => postQueryScreenCounter++}
+      payloadProvider={() =>
+        buildScreenEvent(
+          Screen.POST_QUERY,
+          postQueryScreenCounter,
+          searchSessionId,
+        )
+      }
+    />,
   ];
 }