+++ packages/search/quick-search/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/quick-search
 
+## 2.3.4
+- [patch] Updated dependencies [8a01bcd](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/8a01bcd)
+  - @atlaskit/avatar@12.0.0
+  - @atlaskit/navigation@32.3.1
+  - @atlaskit/item@7.0.7
+
 ## 2.3.3
 - [patch] Replace faker with lightweight internal functions [1c3352a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1c3352a)
 - [none] Updated dependencies [1c3352a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/1c3352a)